MethodArgumentCouldBeFinal - separated java 7 and java 8 tests
