+++ src/test/java/org/walkmod/pmd/ruleset/java/optimizations/visitors/MethodArgumentCouldBeFinalFixtureTest.java
@@ -4,10 +4,19 @@ import org.junit.Test;
 
 import org.walkmod.pmd.ruleset.java.testsupport.SemanticFixtureTest;
 
+import javax.lang.model.SourceVersion;
+
 public class MethodArgumentCouldBeFinalFixtureTest extends SemanticFixtureTest {
 
     @Test
     public void testFixture() throws Exception {
-        doTestFixture(getClass().getSimpleName(), new MethodArgumentCouldBeFinal());
+        doTestFixture(getClass().getSimpleName(), "7", new MethodArgumentCouldBeFinal());
+    }
+
+    @Test
+    public void testFixture8() throws Exception {
+        if (SourceVersion.latestSupported().ordinal() >= 8) {
+            doTestFixture(getClass().getSimpleName(), "8", new MethodArgumentCouldBeFinal());
+        }
     }
 }
