+++ src/test/resources/fixture/pmd/methodargumentcouldbefinal/in/MethodArgumentCouldBeFinalSample7.java
@@ -44,20 +44,4 @@ public class MethodArgumentCouldBeFinalSample {
     public void preDecrementIsNotFinal(int count) {
         --count;
     }
-    
-    public interface F2 {
-        int f(int a, int b);
-    }
-
-    public interface F1 {
-        int f(int a);
-    }
-
-    public void testOnlyLambdaParametersWithTypeCanBeMadeFinal() {
-        F1 o1a = a -> a;
-        F1 o1b = (a) -> a;
-        F1 o1c = (int a) -> a;
-        F2 o2a = (a, b) -> a + b;
-        F2 o2b = (int a, int b) -> a + b;
-    }
 }
