+++ src/test/resources/fixture/pmd/methodargumentcouldbefinal/in/MethodArgumentCouldBeFinalSample8.java
@@ -1,0 +1,18 @@
+public class MethodArgumentCouldBeFinalSample8 {
+
+    public interface F2 {
+        int f(int a, int b);
+    }
+
+    public interface F1 {
+        int f(int a);
+    }
+
+    public void testOnlyLambdaParametersWithTypeCanBeMadeFinal() {
+        F1 o1a = a -> a;
+        F1 o1b = (a) -> a;
+        F1 o1c = (int a) -> a;
+        F2 o2a = (a, b) -> a + b;
+        F2 o2b = (int a, int b) -> a + b;
+    }
+}
